Optionally use IntEnums to set loglvl
Adds functionality to use an IntEnum subclass to set `loglvl`, rather than have to set an actual int.
The IntEnum subclass has the same attributes as slog's logging methods, which makes it easier to set the appropriate level.
The use of IntEnum is effectively an abstraction over the (old) int-style log levels, which makes this feature backwards compatible